Fixed broken pipe message
The logic to detect if a schema already existed was trying to be clever and causing the code to attempt to write to a closed db connection due to the way Go's db connection pool works. The schema detection method has been changed to a simple `show schemas like` query that utilizes an previously created *sql.DB and corrects the go MySQL driver packets.go "broken pipe" error message.